# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: semidx API
  description: |
    Self-hosted semantic code search. The server exposes a REST API under
    `/api/v1`, plus unauthenticated health and metrics endpoints. All request
    and response bodies are JSON unless noted otherwise.

    Two token types are accepted for the `Authorization: Bearer <token>` header:

    - **Opaque API keys** — random `semidx_<hex>` strings (SHA-256 hashed on the
      server; plaintext shown once at creation).
    - **JWT control tokens** — HS256 tokens minted from the `/admin` UI.

    Every authenticated route requires one scope. A token with scope `admin`
    satisfies any requirement.

    | Scope   | Grants                                        |
    |---------|-----------------------------------------------|
    | `read`  | search, list/get projects, get job status     |
    | `write` | create/delete projects, enqueue jobs, push    |
    | `admin` | everything (also used by the bootstrap token) |

    Non-2xx responses carry a JSON error object: `{"error": "human-readable message"}`.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local dev (semidx serve default port)
paths:
...
  /api/v1/projects/{project}/privacy:
    put:
      summary: Set project data-routing policy
      operationId: setProjectPrivacy
      security:
        - bearerAuth: ["write"]
      parameters:
        - name: project
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPrivacyRequest"
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /api/v1/projects:
    post:
      summary: Create a project
      description: |
        Registers a new project. The `source.type` controls how files are
        provided:

        - `push` (default): clients upload files via `files/batch`.
        - `git`: the server clones `source.url` and indexes it; `source.branch`
          defaults to the repository's default branch.

        Request body is limited to 1 MB.
      operationId: createProject
      tags:
        - Projects
      security:
        - bearerAuth: ["write"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
            example:
              name: "myapp"
              model: "bge-m3"
              source:
                type: "git"
                url: "https://github.com/acme/myapp.git"
                branch: "main"
      responses:
        "201":
          description: Project created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid JSON body, invalid project name, or invalid source configuration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Project with this name already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: List projects
      description: Returns all registered projects.
      operationId: listProjects
      tags:
        - Projects
      security:
        - bearerAuth: ["read"]
      responses:
        "200":
          description: A list of projects.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListProjectsResponse"
  /api/v1/projects/{project}:
    get:
      summary: Get a project
      description: Returns a single project by name.
      operationId: getProject
      tags:
        - Projects
      security:
        - bearerAuth: ["read"]
      parameters:
        - name: project
          in: path
          required: true
          schema:
            type: string
          description: URL-escaped project name.
      responses:
        "200":
          description: The project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a project
      description: >
        Removes a project and all its files and chunks (cascade delete).
      operationId: deleteProject
      tags:
        - Projects
      security:
        - bearerAuth: ["write"]
      parameters:
        - name: project
          in: path
          required: true
          schema:
            type: string
          description: URL-escaped project name.
      responses:
        "204":
          description: Project deleted successfully. No response body.
        "404":
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ─── Dependency graph (file↔package) ─────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # also accepts opaque "semidx_<hex>" API keys
      description: >
        Bearer token. Accepted formats: opaque API key (`semidx_<hex>`) or
        HS256 JWT control token. The required scope depends on the endpoint
        (see endpoint descriptions).
  schemas:
...
    Error:
      type: object
      description: A JSON error object returned on non-2xx responses.
      properties:
        error:
          type: string
          description: Human-readable error message.
      required:
        - error

    # ── Project ───────────────────────────────────────────────────────────────
    Project:
      type: object
      description: A registered project.
      properties:
        tenant_id:
          type: integer
        workspace_id:
          type: integer
        name:
          type: string
          description: Project name (unique, alphanumeric + `._-`).
        model:
          type: string
          description: Embedding model name (e.g. `bge-m3`).
        status:
          type: string
          description: Project status (`registered`, `ready`, etc.).
        source_type:
          type: string
          description: Source type (`push` or `git`).
        git_url:
          type: string
          description: Git repository URL (present only when `source_type` is `git`).
        branch:
          type: string
          description: Git branch (present only when `source_type` is `git`).
        privacy_mode:
          type: string
          enum: [cloud, hybrid, edge]
          default: hybrid
          description: Project embedding/data-routing policy.
      required:
        - name
        - model
        - status
        - source_type
...
    ProjectPrivacyRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [cloud, hybrid, edge]
          default: hybrid
      required: [mode]
...
    CreateProjectRequest:
      type: object
      description: Request body for creating a project.
      properties:
        name:
          type: string
          description: >
            Project name. Must start with an alphanumeric character and contain
            only letters, digits, hyphens, underscores, and dots. Max 255 characters.
        model:
          type: string
          description: Embedding model name. Defaults to `bge-m3` if omitted.
          default: bge-m3
        privacy_mode:
          type: string
          enum: [cloud, hybrid, edge]
          default: hybrid
          description: Project embedding/data-routing policy.
        source:
          type: object
          description: Source configuration for the project's files.
          properties:
            type:
              type: string
              enum: [push, git]
              description: >
                How files are provided. `push` (default) means clients upload
                via `files/batch`; `git` means the server clones a repository.
              default: push
            url:
              type: string
              description: >
                Git repository URL. Required when `type` is `git`. Ignored for `push`.
            branch:
              type: string
              description: Git branch (optional; defaults to the repository's default branch).
          required:
            - type
      required:
        - name
    ListProjectsResponse:
      type: object
      description: Response body for listing projects.
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
      required:
        - projects

    # ── Search ────────────────────────────────────────────────────────────────
...